Separating stages: code review generate suggestions and comment suggestions
